refactor: rename error to display-error
