+++ libs/frontend/src/lib/components/error/display-error.svelte
